+++ src/main/java/com/bookspot/batch/BatchServerScheduler.java
@@ -108,34 +108,4 @@ public class BatchServerScheduler {
 
         log.info("Backend 웜업 작업 완료. - {}개의 요청 중 {}개의 요청 실패", urls.size(), failed);
     }
-
-    @Scheduled(cron = "0 1 0 * * MON") // 매주 월요일 00:01
-    public void fetchTop50Weekly() {
-        fetchTop50(RankingType.WEEKLY);
-    }
-
-    @Scheduled(cron = "0 1 0 1 * ?") // 매월 1일 00:01
-    public void fetchTop50Monthly() {
-        fetchTop50(RankingType.MONTHLY);
-    }
-
-    private void fetchTop50(RankingType type) {
-        LocalDate now = LocalDate.now();
-        try {
-            switch (type) {
-                case WEEKLY -> customJobLauncher.launchTop50BooksOfWeek(now);
-                case MONTHLY -> customJobLauncher.launchTop50BooksOfMonth(now);
-            }
-        } catch (Throwable t) {
-            slackAlertService.error(
-                    convertor.convertSimple(
-                            "Top 50 책 스케줄러",
-                            "책 불러오기 실패",
-                            "진행중 예외 발생".concat(t.toString())
-                    )
-            );
-            throw t;
-        }
-    }
-
 }
